// backend/src/common/enums/transaction-status.enum.ts
export enum TransactionStatus {
  PENDIENTE = 'pendiente',
  PENDIENTE_COLOMBIA = 'pendiente_colombia',
  PENDIENTE_VENEZUELA = 'pendiente_venezuela',
  COMPLETADO = 'completado',
  RECHAZADO = 'rechazado',
  CANCELADO_VENDEDOR = 'cancelado_vendedor',
  CANCELADO_ADMINISTRADOR = 'cancelado_administrador',
  TASA_COMPRA_ESTABLECIDA = 'tasa_compra_establecida' // Nuevo estado
}